# Make DB backup use Zstandard compression